last changes

+++ Bycicle.java
@@ -4,7 +4,7 @@ public class Bycicle extends Vehicle {
 	
 	}
 
-	public Bycicle(String sName, double dSpeedMax) {
+	public Bycicle(String sName, float dSpeedMax) {
 		super(sName, dSpeedMax);
 	}
 }
